+++ app/components/user/graphql/GraphQLExplorerDocumentationMutation.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+class GraphQLExplorerDocumentationMutation extends React.PureComponent {
+  render() {
+    return (
+      <div>Zomg Mutation! {this.props.field}</div>
+    )
+  }
+}
+
+export default GraphQLExplorerDocumentationMutation;
